Ported CollisionStage to new design Added debug helper parameter Added directionality pre-condition to check polygon overlap Streamlined boundary functions Fixed actor id to array index map Passing individual buffers from localization Using unsigned int keys in VicinityGrid

+++ TrafficManager/source/pipeline/VicinityGrid.h
@@ -15,19 +15,19 @@ namespace traffic_manager {
     std::shared_timed_mutex modification_mutex;
 
     std::unordered_map<
-      int, std::unordered_set<uint>
+      long, std::unordered_set<uint>
     > grid_to_actor_id;
 
-    std::unordered_map<uint, int> actor_to_grid_id;
+    std::unordered_map<uint, long> actor_to_grid_id;
 
-    int MakeKey(std::pair<int, int> gird_numbers);
+    long MakeKey(std::pair<int, int> gird_numbers);
 
     public:
 
     VicinityGrid();
     ~VicinityGrid();
 
-    std::unordered_set<int> GetActors(carla::SharedPtr<carla::client::Actor> actor);
+    std::unordered_set<uint> GetActors(carla::SharedPtr<carla::client::Actor> actor);
 
   };
 }
